services/categorizer/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
joblib==1.4.0
numpy==1.26.4
protobuf==4.25.3
scikit-learn==1.4.1.post1
scipy==1.13.0
threadpoolctl==3.4.0
